fonts, icons, adds JS, big theme edits
This is essentially the first usable version. It's well tested with
photos and photo sets only. The CSS, JS and fonts are all hosted in my
public Dropbox folder as well, so you can actually just copy and paste
the theme.html file into your custom HTML field of Tumblr.